requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==3.2
django-environ==0.11.2
django-extensions==3.2.3
django-rest-auth==0.9.5
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
ipython==8.12.3
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.18.0
PyJWT==1.7.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing-extensions==4.12.2
wcwidth==0.2.13
